+++ kinto_attachment/views/__init__.py
@@ -3,7 +3,7 @@ import json
 from pyramid import httpexceptions
 from pyramid.settings import asbool
 from kinto.core import logger
-from kinto.core.errors import http_error
+from kinto.core.errors import ERRORS, http_error
 from six import StringIO
 
 from kinto_attachment import utils
@@ -20,10 +20,22 @@ def post_attachment_view(request, file_field):
         # Remove potential existing attachment.
         utils.delete_attachment(request)
 
+    if "multipart/form-data" not in request.headers.get('Content-Type', ''):
+        raise http_error(httpexceptions.HTTPBadRequest(),
+                         errno=ERRORS.INVALID_PARAMETERS,
+                         message="Content-Type should be multipart/form-data")
+
     # Store file locally.
-    content = request.POST.get(file_field)
+    try:
+        content = request.POST.get(file_field)
+    except ValueError as e:
+        raise http_error(httpexceptions.HTTPBadRequest(),
+                         errno=ERRORS.INVALID_PARAMETERS.value,
+                         message=str(e))
+
     if content is None:
         raise http_error(httpexceptions.HTTPBadRequest(),
+                         errno=ERRORS.INVALID_POSTED_DATA,
                          message="Attachment missing.")
 
     randomize = True